160427
전체 변경 : 탭을 개행 4개로 변경
완료 : 01550, 02743, 05337
진행 : 02178

+++ ACMICPC/ACMICPC/01000-02000/01032.c
@@ -4,31 +4,31 @@
 
 int Problem01032(void)
 {
-	char arr_strFileName[50][51] = { 0, };
-	char strResult[51] = { 0, };
-	int nFileNum = 0;
-	int nFileNameSize = 0;
+    char arr_strFileName[50][51] = { 0, };
+    char strResult[51] = { 0, };
+    int nFileNum = 0;
+    size_t nFileNameSize = 0;
 
-	scanf("%d", &nFileNum);
-	for (int i = 0; i < nFileNum; i++)
-	{
-		scanf("%s", arr_strFileName[i]);
-	}
+    scanf("%d", &nFileNum);
+    for (int i = 0; i < nFileNum; i++)
+    {
+        scanf("%s", arr_strFileName[i]);
+    }
 
-	nFileNameSize = strlen(arr_strFileName[0]);
-	strcpy(strResult, arr_strFileName[0]);
+    nFileNameSize = strlen(arr_strFileName[0]);
+    strcpy(strResult, arr_strFileName[0]);
 
-	for (int i = 0; i < nFileNameSize; i++)
-	{
-		for (int j = 1; j < nFileNum; j++)
-		{
-			if (strResult[i] != arr_strFileName[j][i])
-			{
-				strResult[i] = '?';
-				break;
-			}
-		}
-	}
-	printf("%s\n", strResult);
-	return 0;
+    for (int i = 0; i < nFileNameSize; i++)
+    {
+        for (int j = 1; j < nFileNum; j++)
+        {
+            if (strResult[i] != arr_strFileName[j][i])
+            {
+                strResult[i] = '?';
+                break;
+            }
+        }
+    }
+    printf("%s\n", strResult);
+    return 0;
 }
